refactor modules to work for native similarly as web
