+++ src/client/app/routes.web.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Switch } from 'react-router-dom';
+
+import modules from '../modules';
+
+export default (
+  <Switch>
+    {modules.routes}
+  </Switch>
+);
