+++ src/client/modules/common/components/index.js
@@ -1,3 +1,4 @@
 // eslint-disable-next-line import/prefer-default-export
 export { default as Button } from './button';
 export { default as RenderField } from './render_field';
+export { default as createTabBarIconWrapper } from './tab_bar_icon';
