+++ src/client/modules/common/components/tab_bar_icon.js
@@ -1,0 +1,10 @@
+/* eslint-disable react/display-name */
+import React from 'react';
+
+const createTabBarIconWrapper = (
+  TabBarIconComponent,
+  defaultProps
+  // eslint-disable-next-line react/prop-types
+) => props => <TabBarIconComponent {...defaultProps} color={props.tintColor} />;
+
+export default createTabBarIconWrapper;
