+++ src/client/modules/counter/index.web.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Route } from 'react-router-dom';
+
+import Counter from './containers/counter';
+import reducers from './reducers';
+
+import Feature from '../connector';
+
+export default new Feature({
+  route: <Route exact path="/" component={Counter}/>,
+  reducer: { counter: reducers }
+});
